💄 improve navigation
