+++ apps/atom.io.fyi/src/app/docs/Toggle.tsx
@@ -1,0 +1,9 @@
+import { Button } from "./Toggle.Button"
+
+export type ToggleProps = {
+	children: string
+	checked: boolean
+	onChange: () => void
+}
+
+export const Toggle = { Button }
