+++ apps/atom.io.fyi/src/app/globals.scss
@@ -78,6 +78,7 @@ h3 {
   --bg-soft-3: var(--bg-tint-3);
   --bg-soft-glass-1: #333a;
   --bg-soft-glass-2: #333c;
+  --bg-soft-glass-3: #333e;
   --hyperlink-color: #09f;
   --hyperlink-hover-color: #0df;
   --hyperlink-visited-color: #99f;
@@ -106,6 +107,7 @@ h3 {
     --bg-soft-3: var(--bg-shade-3);
     --bg-soft-glass-1: #fff8;
     --bg-soft-glass-2: #fffa;
+    --bg-soft-glass-3: #fffc;
 
     --hyperlink-color: #08f;
     --hyperlink-hover-color: #04f;
